requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.15.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pycparser==2.22
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
